RetentionConfig: restrict days to uint16
As discussed in #821, allowing huge retention values - as the type
uint64 allows - may result in overflows. Especially Go's time.AddDate()
method silently overflows for huge values, resulting in unexpected
comparison times.

> $ ./icingadb --config <(echo 'retention: {history-days: -1}')
> can't parse YAML file /proc/self/fd/11: cannot unmarshal -1 into Go value of type uint16 ( overflow )
>
> $ ./icingadb --config <(echo 'retention: {history-days: 100000}')
> can't parse YAML file /proc/self/fd/11: cannot unmarshal 100000 into Go value of type uint16 ( overflow )

+++ internal/config/config.go
@@ -55,8 +55,8 @@ type Flags struct {
 
 // RetentionConfig defines configuration for history retention.
 type RetentionConfig struct {
-	HistoryDays uint64                   `yaml:"history-days"`
-	SlaDays     uint64                   `yaml:"sla-days"`
+	HistoryDays uint16                   `yaml:"history-days"`
+	SlaDays     uint16                   `yaml:"sla-days"`
 	Interval    time.Duration            `yaml:"interval" default:"1h"`
 	Count       uint64                   `yaml:"count" default:"5000"`
 	Options     history.RetentionOptions `yaml:"options"`
